refactor: 패키지 구조 애그리거트에 맞게 변경
