+++ src/main/java/com/atwoz/mission/domain/mission/vo/MissionType.java
@@ -1,4 +1,4 @@
-package com.atwoz.mission.domain.vo;
+package com.atwoz.mission.domain.mission.vo;
 
 import com.atwoz.mission.exception.MissionTypeInvalidException;
 import lombok.Getter;
