+++ src/test/java/com/atwoz/mission/domain/membermission/MemberMissionsTest.java
@@ -1,5 +1,7 @@
-package com.atwoz.mission.domain;
+package com.atwoz.mission.domain.membermission;
 
+import com.atwoz.mission.domain.membermission.MemberMission;
+import com.atwoz.mission.domain.membermission.MemberMissions;
 import com.atwoz.mission.exception.MissionNotFoundException;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
