+++ src/test/java/com/atwoz/mission/domain/mission/MissionTest.java
@@ -1,7 +1,8 @@
-package com.atwoz.mission.domain;
+package com.atwoz.mission.domain.mission;
 
-import com.atwoz.mission.domain.vo.MissionType;
-import com.atwoz.mission.domain.vo.PublicOption;
+import com.atwoz.mission.domain.mission.Mission;
+import com.atwoz.mission.domain.mission.vo.MissionType;
+import com.atwoz.mission.domain.mission.vo.PublicOption;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
