+++ src/test/java/com/atwoz/mission/domain/mission/vo/PublicOptionTest.java
@@ -1,4 +1,4 @@
-package com.atwoz.mission.domain.vo;
+package com.atwoz.mission.domain.mission.vo;
 
 import com.atwoz.mission.exception.PublicOptionInvalidException;
 import org.junit.jupiter.api.DisplayNameGeneration;
